// src/app/models/evt-models.ts
export interface HighlightData {
  highlight: boolean;
  highlightColor: string;
}

export class GenericElement {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  type: Type<any>;
  path?: string;
  class?: string;
  attributes: Attributes;
  content: Array<ParseResult<GenericElement>>;
}
// ...
export interface Attributes {
  [key: string]: string;
}
// ...
export type PlacementType =
  | 'above'
  | 'below'
  | 'inline'
  | 'left'
  | 'right'
  | 'inspace'
  | 'end'
  | 'sup'
  | 'sub'
  | 'under';

export class Addition extends GenericElement {
  place: PlacementType;
}